requirements.txt
Django==1.5.3
South==0.8.2
dj-database-url==0.2.2
dj-static==0.0.5
django-crispy-forms==1.4.0
django-toolbelt==0.0.1
factory-boy==2.1.2
gunicorn==18.0
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
